requirements.txt
langgraph==0.6.6
langchain==0.3.27
langchain-community
langchain-core
langchain_aws
watchdog

pdfplumber
python-docx
pillow


fastapi
uvicorn
pydantic
pyyaml

faiss-cpu
sentence-transformers

streamlit

watchdog
requests
pytesseract